Split weights.h into a source and header file
